+++ src/main/java/simpaths/model/decisions/States.java
@@ -890,7 +890,7 @@ public class States {
     SocialCareReceiptState getSocialCareReceiptStateCode() {
         SocialCareReceiptState code;
         if (Parameters.flagSocialCare && ageYears >= DecisionParams.minAgeReceiveFormalCare)
-            code = SocialCareReceiptState.getCode((int)getVal(Axis.SocialCareReceiptState));
+            code = SocialCareReceiptState.getCode(getVal(Axis.SocialCareReceiptState));
         else
             code = SocialCareReceiptState.NoneNeeded;
         return code;
@@ -902,7 +902,7 @@ public class States {
     SocialCareReceipt getSocialCareReceiptCode() {
         SocialCareReceipt code;
         if (Parameters.flagSocialCare && ageYears >= DecisionParams.minAgeReceiveFormalCare)
-            code = SocialCareReceipt.getCode((int)getVal(Axis.SocialCareReceiptState));
+            code = SocialCareReceipt.getCode(getVal(Axis.SocialCareReceiptState));
         else
             code = SocialCareReceipt.None;
         return code;
@@ -914,7 +914,7 @@ public class States {
     SocialCareProvision getSocialCareProvisionCode() {
         SocialCareProvision code;
         if (Parameters.flagSocialCare)
-            code = SocialCareProvision.getCode((int)getVal(Axis.SocialCareProvision));
+            code = SocialCareProvision.getCode(getVal(Axis.SocialCareProvision));
         else
             code = SocialCareProvision.None;
         return code;
